requirements.txt
pymongo 
openai 
python-dotenv
tiktoken
langchain-openai
azure-storage-blob
azure-identity
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
fastapi
uvicorn
weaviate-client==3.24.1
langchain-mongodb